+++ lib/webview_macos.dart
@@ -1,3 +1,5 @@
+import 'package:flutter/material.dart';
+
 import 'webview_macos_platform_interface.dart';
 
 class WebviewMacos {
@@ -5,7 +7,27 @@ class WebviewMacos {
     return WebviewMacosPlatform.instance.showWebView(initialURL, reset);
   }
 
-  Future<void> loadURL(String url) {
+  static Future<void> showWebViewWithArgs({
+    required String url,
+    bool resetPreviousInstance = true,
+    String windowTitle = "WebView",
+    Function(String, String, FlutterError?)? onNavigationStart,
+    Function(String, String, FlutterError?)? onNavigationCommit,
+    Function(String, String, FlutterError?)? onNavigationError,
+    Function(String, String, FlutterError?)? onNavigationFinished,
+  }) async {
+    return WebviewMacosPlatform.instance.showWebViewWithArgs(
+      url: url,
+      resetPreviousInstance: resetPreviousInstance,
+      windowTitle: windowTitle,
+      onNavigationStart: onNavigationStart,
+      onNavigationCommit: onNavigationCommit,
+      onNavigationError: onNavigationError,
+      onNavigationFinished: onNavigationFinished,
+    );
+  }
+
+  static Future<void> loadURL(String url) {
     return WebviewMacosPlatform.instance.loadURL(url);
   }
 
@@ -17,7 +39,7 @@ class WebviewMacos {
     return WebviewMacosPlatform.instance.evaluateJavaScript(jsString);
   }
 
-  Future<void> didFinish(Function(String, String, bool) didFinish) {
+  Future<void> didFinish(Function(String, String, FlutterError?) didFinish) {
     return WebviewMacosPlatform.instance.didFinish(didFinish);
   }