+++ lib/webview_macos_method_channel.dart
@@ -11,12 +11,48 @@ class MethodChannelWebviewMacos extends WebviewMacosPlatform {
 
   bool methodCallHandlerHasBeenSet = false;
   Function(String, String, bool)? didFinishCallback;
+  Map<String, dynamic Function(String, String, FlutterError?)?> savedCallbacks =
+      {};
 
   @override
+  @deprecated
   Future<void> showWebView(String initialURL, [bool reset = true]) async {
     final _ =
         await methodChannel.invokeMethod('showWebView', [initialURL, reset]);
     return;
+  }
+
+  @override
+  Future<void> showWebViewWithArgs({
+    required String url,
+    bool resetPreviousInstance = true,
+    String windowTitle = "WebView",
+    double? windowWidth,
+    double? windowHeight,
+    Function(String, String, FlutterError?)? onNavigationStart,
+    Function(String, String, FlutterError?)? onNavigationCommit,
+    Function(String, String, FlutterError?)? onNavigationError,
+    Function(String, String, FlutterError?)? onNavigationFinished,
+  }) {
+    savedCallbacks.addAll({
+      "onNavigationStart": onNavigationStart,
+      "onNavigationCommit": onNavigationCommit,
+      "onNavigationError": onNavigationError,
+      "onNavigationFinished": onNavigationFinished,
+    });
+    if (!methodCallHandlerHasBeenSet) {
+      methodChannel.setMethodCallHandler(this._genericMethodHandler);
+      methodCallHandlerHasBeenSet = true;
+    }
+    return methodChannel.invokeMethod("showWebViewWithArgs", {
+      "url": url,
+      "windowTitle": windowTitle,
+      "resetPreviousInstance": true,
+      "onNavigationStart": onNavigationStart != null,
+      "onNavigationCommit": onNavigationCommit != null,
+      "onNavigationError": onNavigationError != null,
+      "onNavigationFinished": onNavigationFinished != null,
+    });
   }
 
   @override
@@ -40,12 +76,9 @@ class MethodChannelWebviewMacos extends WebviewMacosPlatform {
   }
 
   @override
-  Future<void> didFinish(Function(String, String, bool) didFinish) async {
-    if (!methodCallHandlerHasBeenSet) {
-      methodChannel.setMethodCallHandler(this._genericMethodHandler);
-      methodCallHandlerHasBeenSet = true;
-    }
-    didFinishCallback = didFinish;
+  Future<void> didFinish(
+      Function(String, String, FlutterError?) didFinish) async {
+    savedCallbacks["onNavigationFinish"] = didFinish;
     final bool _ = await methodChannel.invokeMethod("didFinish") ?? false;
     return;
   }
@@ -58,8 +91,8 @@ class MethodChannelWebviewMacos extends WebviewMacosPlatform {
 
   Future<void> _genericMethodHandler(MethodCall call) async {
     switch (call.method) {
-      case "didFinish":
-        if (didFinishCallback != null) {
+      case "onNavigationError":
+        if (savedCallbacks["onNavigationError"] != null) {
           String resp = "";
           String url = "";
           if (call.arguments is List<dynamic>) {