+++ macos/Classes/CGSize+Extension.swift
@@ -1,0 +1,14 @@
+//
+//  CGSize+Extension.swift
+//  webview_macos
+//
+//  Created by riccardo on 01/11/22.
+//
+
+import Foundation
+
+extension CGSize {
+    static var defaultWindowSize: CGSize {
+        CGSize(width: 1000, height: 600)
+    }
+}