+++ macos/Classes/NSViewController+Extension.swift
@@ -9,17 +9,16 @@ import Foundation
 
 extension NSViewController {
     
-    func presentInNewWindow(viewController: NSViewController, title: String = "") -> NSWindowController {
+    func presentInNewWindow(viewController: NSViewController, windowTitle: String = "", windowSize: CGSize = CGSize(width: 1000, height: 600)) -> NSWindowController {
         let window = NSWindow(contentViewController: viewController)
         
         var rect = window.contentRect(forFrameRect: window.frame)
-        // Set your frame width here
-        rect.size = .init(width: 1000, height: 600)
+        rect.size = windowSize
         let frame = window.frameRect(forContentRect: rect)
         window.setFrame(frame, display: true, animate: true)
         
         window.makeKeyAndOrderFront(self)
-        window.title = title
+        window.title = windowTitle
         let windowVC = NSWindowController(window: window)
         windowVC.showWindow(self)