+++ test/webview_macos_test.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/src/foundation/assertions.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:webview_macos/webview_macos_method_channel.dart';
 import 'package:webview_macos/webview_macos_platform_interface.dart';
@@ -32,7 +33,21 @@ class MockNewWebviewMacosPlatform
   }
 
   @override
-  Future<void> didFinish(Function(String, String, bool) didFinish) {
+  Future<void> didFinish(Function(String, String, FlutterError?) didFinish) {
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<void> showWebViewWithArgs(
+      {required String url,
+      bool resetPreviousInstance = true,
+      String windowTitle = "",
+      Function(String p1, String p2, FlutterError? p3)? onNavigationStart,
+      Function(String p1, String p2, FlutterError? p3)? onNavigationCommit,
+      Function(String p1, String p2, FlutterError? p3)? onNavigationError,
+      Function(String p1, String p2, FlutterError? p3)? onNavigationFinished,
+      double? windowWidth,
+      double? windowHeight}) {
     throw UnimplementedError();
   }
 }