+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+## 0.0.3
+
+* Changes on Navigation Policies
+
 ## 0.0.2
 
 * Added reset flag on showWebView to let user reset the webview everytime it's opened 
 
 ## 0.0.1
 
-* Stub ok WKWebView functions completed
+* Stub of WKWebView functions completed